requirements.txt
asgiref==3.3.1
autopep8==1.5.5
cffi==1.14.4
cryptography==3.3.1
Django==3.1.6
django-rest-knox==4.1.0
djangorestframework==3.12.2
pycodestyle==2.6.0
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
